requirements.txt
torch>=2.0.0,<2.4.0
torchvision>=0.15.0,<0.19.0
transformers>=4.30.0,<4.45.0
pillow>=10.0.0
chromadb>=0.4.0
numpy>=1.24.0
tqdm>=4.65.0
protobuf>=3.20.0,<4.25.0
safetensors>=0.4.0
flask>=2.3.0
flask-cors>=4.0.0

